+++ src/YarnSpinner.Console/AssemblyInfo.cs
@@ -1,4 +1,7 @@
 ﻿using System.Reflection;
 
 // During CI builds,this file will be populated by GitVersion with version information
-// at build time.
+// at build time, and these placeholder version numbers will be replaced.
+[assembly: AssemblyVersion("0.0.0.0")]
+[assembly: AssemblyInformationalVersion("0.0.0+0.Branch.unknown.Sha.0000000000000000000000000000000000000000")]
+[assembly: AssemblyFileVersion("0.0.0.0")]