+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- New subcommand `extract` allows for extraction of lines for the purpose of VO.
+
 ### Changed
 
 ### Removed